moved article components to separate article-components directory and updated all mappings
