+++ client/kbarticle-interface/src/app/Articles-module/components/article-components/display-article/display-article.component.scss
